Update README.md
Added some things that weren't obvious to me the first time I tried to use the package.

I also adjusted the indentation so it's clearer that the IndexRoute is actually wrapped by the RequireAuthContainer route. I'm not sure whether all routes below the container will be authorized as well, so feel free to fix that wording if it's not correct.